+++ synapse/tests/test_lib_trigger.py
@@ -125,22 +125,26 @@ class TrigTest(s_t_utils.SynTest):
                 await self.agenlen(1, core.eval('test:int=8'))
 
                 # Bad trigger parms
-                await self.asyncraises(s_exc.BadOptValu, core.addTrigger('nocond', 'test:int=4',
-                                                                         info={'form': 'test:str'}))
-                await self.asyncraises(s_exc.BadSyntax,
-                                       core.addTrigger('node:add', ' | | badstorm ', info={'form': 'test:str'}))
-                await self.asyncraises(s_exc.BadOptValu,
-                                       core.addTrigger('node:add', 'test:int=4', info={'form': 'test:str', 'tag': 'foo'}))
-                await self.asyncraises(s_exc.BadOptValu,
-                                       core.addTrigger('prop:set', 'test:int=4',
-                                                       info={'form': 'test:str', 'prop': 'foo'}))
-                await self.asyncraises(s_exc.BadOptValu,
-                                       core.addTrigger('tag:add', '[ +#count test:str=$tag ]', info={}))
-                await self.asyncraises(s_exc.BadOptValu, core.addTrigger('tag:add', '[ +#count test:str=$tag ]',
-                                                                         info={'tag': 'foo', 'prop': 'test:str'}))
+                with self.raises(s_exc.BadOptValu):
+                    await core.addTrigger('nocond', 'test:int=4', info={'form': 'test:str'})
+
+                with self.raises(s_exc.BadSyntax):
+                    await core.addTrigger('node:add', ' | | badstorm ', info={'form': 'test:str'})
+
+                with self.raises(s_exc.BadOptValu):
+                    await core.addTrigger('node:add', 'test:int=4', info={'form': 'test:str', 'tag': 'foo'})
+
+                with self.raises(s_exc.BadOptValu):
+                    await core.addTrigger('prop:set', 'test:int=4', info={'form': 'test:str', 'prop': 'foo'})
+
+                with self.raises(s_exc.BadOptValu):
+                    await core.addTrigger('tag:add', '[ +#count test:str=$tag ]', info={})
+
+                with self.raises(s_exc.BadOptValu):
+                    await core.addTrigger('tag:add', '[ +#count test:str=$tag ]', info={'tag': 'foo', 'prop': 'test:str'})
                 # bad tagmatch
-                await self.asyncraises(s_exc.BadTag,
-                                       core.addTrigger('tag:add', '[ +#count test:str=$tag ]', info={'tag': 'foo&baz'}))
+                with self.raises(s_exc.BadTag):
+                    await core.addTrigger('tag:add', '[ +#count test:str=$tag ]', info={'tag': 'foo&baz'})
 
                 # Trigger list
                 triglist = await core.listTriggers()
@@ -176,15 +180,19 @@ class TrigTest(s_t_utils.SynTest):
                 buid2 = [b for b, r in triglist if r['cond'] == 'tag:add' and r.get('form') is None][0]
                 await core.delTrigger(buid2)
 
+                # A rando user can't manipulate triggers
+
                 await core.addAuthUser('fred')
                 await core.setUserPasswd('fred', 'fred')
 
                 url = real.getLocalUrl()
 
                 async with await s_telepath.openurl(url, user='fred') as fred:
-
                     # Trigger list other user
                     self.len(0, await fred.listTriggers())
+
+                    with self.raises(s_exc.AuthDeny):
+                        await fred.addTrigger('node:add', '[ test:int=1 ]', info={'form': 'test:str'})
 
                     # Delete trigger auth failure
                     await self.asyncraises(s_exc.AuthDeny, fred.delTrigger(buid))
@@ -245,3 +253,44 @@ class TrigTest(s_t_utils.SynTest):
 
             self.len(0, await core.eval('#count0').list())
             self.len(0, await core.eval('#count1').list())
+
+    async def test_trigger_perms(self):
+
+        async with self.getTestCore() as core:
+
+            visi = await core.auth.addUser('visi')
+            newb = await core.auth.addUser('newb')
+
+            await visi.addRule((True, ('tag:add', 'foo')))
+
+            async with core.getLocalProxy(user='visi') as proxy:
+
+                with self.raises(s_exc.AuthDeny):
+                    await proxy.addTrigger('node:add', '[ +#foo ]', info={'form': 'inet:ipv4'})
+
+                await visi.addRule((True, ('trigger', 'add')))
+
+                trig0 = await proxy.addTrigger('node:add', '[ +#foo ]', info={'form': 'inet:ipv4'})
+                trig1 = await proxy.addTrigger('node:add', '[ +#foo ]', info={'form': 'inet:ipv6'})
+
+                nodes = await core.nodes('[ inet:ipv4=1.2.3.4 ]')
+                self.nn(nodes[0].tags.get('foo'))
+
+                await proxy.delTrigger(trig0)
+
+            async with core.getLocalProxy(user='newb') as proxy:
+
+                with self.raises(s_exc.AuthDeny):
+                    await proxy.delTrigger(trig1)
+
+                self.eq(await proxy.listTriggers(), ())
+                await newb.addRule((True, ('trigger', 'get')))
+                self.len(1, await proxy.listTriggers())
+
+                with self.raises(s_exc.AuthDeny):
+                    await proxy.disableTrigger(trig1)
+                await newb.addRule((True, ('trigger', 'set')))
+                self.none(await proxy.disableTrigger(trig1))
+
+                await newb.addRule((True, ('trigger', 'del')))
+                await proxy.delTrigger(trig1)