+++ synapse/tests/test_telepath.py
@@ -172,12 +172,25 @@ class TeleTest(s_t_utils.SynTest):
 
         async with self.getTestDmon() as dmon:
 
-            addr = await dmon.listen('tcp://127.0.0.1:0')
             dmon.share('foo', foo)
 
             await self.asyncraises(s_exc.BadUrl, s_telepath.openurl('noscheme/foo'))
 
-            prox = await s_telepath.openurl('tcp://127.0.0.1/foo', port=addr[1])
+            prox = await s_telepath.openurl('tcp://127.0.0.1/foo', port=dmon.addr[1])
+
+            # Some bookkeeping data about the connection is available
+            # from the daemon related to the session's objects and
+            # connection information.
+            snfo = await dmon.getSessInfo()
+            self.len(1, snfo)
+            self.eq(snfo[0].get('items'), {None: 'synapse.tests.test_telepath.Foo'})
+            conninfo = snfo[0].get('conninfo')
+            self.isinstance(conninfo, dict)
+            self.eq(conninfo.get('family'), 'tcp')
+            self.eq(conninfo.get('ipver'), 'ipv4')
+            # The prox's local sock.getsockname() corresponds to the
+            # server's sock.getpeername()
+            self.eq(conninfo.get('addr'), prox.link.sock.getsockname())
 
             # Prox exposes remote synapse version
             self.eq(prox._getSynVers(), s_version.version)
@@ -229,10 +242,9 @@ class TeleTest(s_t_utils.SynTest):
 
         async with self.getTestDmon() as dmon:
 
-            addr = await dmon.listen('tcp://127.0.0.1:0')
             dmon.share('foo', foo)
 
-            async with await s_telepath.openurl('tcp://127.0.0.1/foo', port=addr[1]) as prox:
+            async with await s_telepath.openurl('tcp://127.0.0.1/foo', port=dmon.addr[1]) as prox:
                 self.eq((10, 20, 30), await s_coro.executor(sync))
 
     def test_telepath_sync_genr_break(self):
@@ -280,12 +292,11 @@ class TeleTest(s_t_utils.SynTest):
 
         async with self.getTestDmon() as dmon:
 
-            addr = await dmon.listen('tcp://127.0.0.1:0')
             dmon.share('foo', foo)
 
             await self.asyncraises(s_exc.BadUrl, s_telepath.openurl('noscheme/foo'))
 
-            async with await s_telepath.openurl('tcp://127.0.0.1/foo', port=addr[1]) as prox:
+            async with await s_telepath.openurl('tcp://127.0.0.1/foo', port=dmon.addr[1]) as prox:
 
                 prox.sess = None
 
@@ -356,15 +367,20 @@ class TeleTest(s_t_utils.SynTest):
             async with await s_telepath.openurl(f'ssl://{hostname}/foo', port=addr[1]) as prox:
                 self.eq(30, await prox.bar(10, 20))
 
+                # The daemon's session information for a TLS link
+                # its own family.
+                sessions = await dmon.getSessInfo()
+                self.len(1, sessions)
+                self.eq(sessions[0].get('conninfo').get('family'), 'tls')
+
     async def test_telepath_surrogate(self):
 
         foo = Foo()
         async with self.getTestDmon() as dmon:
 
-            addr = await dmon.listen('tcp://127.0.0.1:0')
             dmon.share('foo', foo)
 
-            async with await s_telepath.openurl('tcp://127.0.0.1/foo', port=addr[1]) as prox:
+            async with await s_telepath.openurl('tcp://127.0.0.1/foo', port=dmon.addr[1]) as prox:
                 bads = '\u01cb\ufffd\ud842\ufffd\u0012'
                 t0 = ('1234', {'key': bads})
 
@@ -378,10 +394,9 @@ class TeleTest(s_t_utils.SynTest):
 
         async with self.getTestDmon() as dmon:
 
-            addr = await dmon.listen('tcp://127.0.0.1:0')
             dmon.share('foo', foo)
 
-            async with await s_telepath.openurl('tcp://127.0.0.1/foo', port=addr[1]) as prox:
+            async with await s_telepath.openurl('tcp://127.0.0.1/foo', port=dmon.addr[1]) as prox:
 
                 genr = prox.corogenr(3)
                 self.eq([0, 1, 2], [x async for x in genr])
@@ -461,6 +476,10 @@ class TeleTest(s_t_utils.SynTest):
                 # Ensure the session tracks the reference to the TeleApi object
                 sess = dmon.sessions[list(dmon.sessions.keys())[0]]
                 self.isinstance(sess.getSessItem(None), TeleApi)
+                # And that data is available from the session helper API
+                snfo = await dmon.getSessInfo()
+                self.len(1, snfo)
+                self.eq(snfo[0].get('items'), {None: 'synapse.tests.test_telepath.TeleApi'})
 
                 self.eq(10, await proxy.getFooBar(20, 10))
 
@@ -489,6 +508,14 @@ class TeleTest(s_t_utils.SynTest):
                 self.true(await asyncio.wait_for(evt.wait(), 6))
                 self.len(2, sess.items)
                 self.nn(sess.getSessItem(key))
+
+                # ensure that the share is represented in the session info via
+                # the helper APIs as well
+                snfo = await dmon.getSessInfo()
+                self.len(1, snfo)
+                self.eq(snfo[0].get('items'),
+                        {None: 'synapse.tests.test_telepath.TeleApi',
+                         key: 'synapse.tests.test_telepath.CustomShare'})
 
                 # and we can still use the first obj we made
                 ret = await alist(obj.custgenr(3))
@@ -532,12 +559,12 @@ class TeleTest(s_t_utils.SynTest):
 
         async with self.getTestDmon() as dmon:
 
-            addr = await dmon.listen('tcp://127.0.0.1:0')
+            host, port = dmon.addr
             dmon.share(name, item)
 
             with self.getTestDir() as dirn:
 
-                url = f'tcp://{addr[0]}:{addr[1]}/{name}'
+                url = f'tcp://{host}:{port}/{name}'
                 beepbeep_alias = url + '/beepbeep'
                 aliases = {name: url,
                            f'{name}/borp': beepbeep_alias}
@@ -569,10 +596,10 @@ class TeleTest(s_t_utils.SynTest):
 
         async with self.getTestDmon() as dmon:
 
-            addr, port = await dmon.listen('tcp://127.0.0.1:0')
+            host, port = dmon.addr
             dmon.share('*', Foo())
 
-            async with await s_telepath.openurl(f'tcp://{addr}:{port}/') as prox:
+            async with await s_telepath.openurl(f'tcp://{host}:{port}/') as prox:
                 self.eq('hiya', await prox.echo('hiya'))
 
     async def test_url_cell(self):
@@ -580,16 +607,59 @@ class TeleTest(s_t_utils.SynTest):
         with self.getTestDir(chdir=True) as dirn:
 
             path = os.path.join(dirn, 'cell')
+            sockpath = os.path.join(path, 'sock')
 
             async with await s_cell.Cell.anit(path) as cell:
 
                 # test a relative cell:// url
                 async with await s_telepath.openurl('cell://cell') as prox:
                     self.eq('cell', await prox.getCellType())
+                    # unix path information is available from the session information.
+                    snfo = await cell.dmon.getSessInfo()
+                    self.eq(snfo[0].get('conninfo'),
+                            {'family': 'unix',
+                             'addr': sockpath})
 
                 # test an absolute cell:// url
                 async with await s_telepath.openurl(f'cell://{path}') as prox:
                     self.eq('cell', await prox.getCellType())
+                    # unix path information is available from the session information.
+                    snfo = await cell.dmon.getSessInfo()
+                    self.eq(snfo[0].get('conninfo'),
+                            {'family': 'unix',
+                             'addr': sockpath})
+
+    async def test_ipv6(self):
+
+        foo = Foo()
+
+        async with self.getTestDmon() as dmon:
+
+            dmon.share('foo', foo)
+            try:
+                addr = await dmon.listen('tcp://[::1]:0/')
+            except asyncio.CancelledError:
+                raise
+            except OSError:
+                if os.getenv('CIRCLECI', False):
+                    # Circleci container tests do not support IPV6 (but osx does)
+                    # https://circleci.com/docs/2.0/faq/#can-i-use-ipv6-in-my-tests
+                    self.skip('ipv6 is not supported in circleci')
+                else:
+                    raise
+            host, port = addr[0], addr[1]
+
+            async with await s_telepath.openurl(f'tcp://{host}/foo',
+                                                port=port) as prox:
+                # Ensure that ipv6 is returned via session info
+                snfo = await dmon.getSessInfo()
+                conninfo = snfo[0].get('conninfo')
+                self.eq(conninfo, {'family': 'tcp',
+                               'ipver': 'ipv6',
+                               'addr': prox.link.sock.getsockname()})
+
+                # check a standard return value
+                self.eq(30, await prox.bar(10, 20))
 
     async def test_telepath_client_redir(self):
 
@@ -703,3 +773,76 @@ class TeleTest(s_t_utils.SynTest):
             self.eq(2, fail1.count)
 
         await dmon1.fini()
+
+    async def test_telepath_poolsize(self):
+
+        # While test_telepath_sync_genr_break also touches the link pool,
+        # it doesn't validate the pool size or automatic link teardown
+        # behavior when a extra link is placed into the pool.
+        foo = Foo()
+
+        async with self.getTestDmon() as dmon:
+            dmon.share('foo', foo)
+            url = f'tcp://127.0.0.1:{dmon.addr[1]}/foo'
+
+            # Validate the Proxy behavior then the client override
+            prox = await s_telepath.openurl(url)  # type: Foo
+            prox._link_poolsize = 2
+
+            # Start with no links
+            self.len(0, prox.links)
+            self.eq(await prox.echo(1), 1)
+
+            # We now have one link - spin up a generator to grab it
+            self.len(1, prox.links)
+            l0 = prox.links[0]
+            genr = await prox.genr()  # type: s_coro.GenrHelp
+            self.eq(await genr.genr.__anext__(), 10)
+
+            # The link is being used by the genr
+            self.len(0, prox.links)
+
+            # and upon exhuastion, that link is put back
+            self.eq(await genr.list(), (20, 30))
+            self.len(1, prox.links)
+            self.true(prox.links[0] is l0)
+
+            # Grab the existing link, then do two more calls
+            genr0 = await prox.genr()  # contains l0
+            genr1 = await prox.genr()
+            genr2 = await prox.genr()
+            self.len(0, prox.links)
+            # Consume two of the three generators
+            self.eq(await genr2.list(), (10, 20, 30))
+            self.len(1, prox.links)
+            self.eq(await genr1.list(), (10, 20, 30))
+            self.len(2, prox.links)
+            # Exhausting the lsat generator results in his
+            # link not being placed back into the pool
+            self.eq(await genr0.list(), (10, 20, 30))
+            self.len(2, prox.links)
+            links = set(l for l in prox.links)
+            self.notin(l0, links)
+            # And that link l0 has been fini'd
+            self.true(l0.isfini)
+
+            # Tear down a link by hand and place it back
+            # into the pool - that will fail b/c the link
+            # has been down down.
+            l1 = await prox.getPoolLink()
+            self.len(1, prox.links)
+            await l1.fini()
+            await prox._putPoolLink(l1)
+            self.len(1, prox.links)
+
+            # And all our links are torn down on fini
+            await prox.fini()
+            self.len(1, prox.links)
+            for link in prox.links:
+                self.true(link.isfini)
+
+            # The telepath Client passes through this value as a configuration parameter
+            conf = {'link_poolsize': 2, 'timeout': 2}
+            async with await s_telepath.Client.anit(url, conf=conf) as client:
+                await client.waitready()
+                self.true(client._t_proxy._link_poolsize, 2)